expose diff
